requirements.txt
appdirs==1.4.3
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
factory-boy==2.8.1
Faker==0.7.11
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
